t pull

+++ framework/Samplers/DynamicEventTree.py
@@ -21,14 +21,13 @@ import abc
 import numpy as np
 import json
 from operator import mul,itemgetter
-from collections import OrderedDict
 from functools import reduce
 from scipy import spatial
 from scipy.interpolate import InterpolatedUnivariateSpline
 import xml.etree.ElementTree as ET
 import itertools
 from math import ceil
-from collections import OrderedDict
+from collections import OrderedDict, Counter
 from sklearn import neighbors
 from sklearn.utils.extmath import cartesian
 
@@ -80,6 +79,8 @@ class DynamicEventTree(Grid):
     # print the xml tree representation of the dynamic event tree calculation
     # see variable 'self.TreeInfo'
     self.printEndXmlSummary                = False
+    # flag to control if the branch info xml file needs to be removed after reading
+    self.removeXmlBranchInfo               = True
     # Dictionary of the probability bins for each distribution that have been
     #  inputted by the user ('distName':[Pb_Threshold_1, Pb_Threshold_2, ..., Pb_Threshold_n])
     self.branchProbabilities               = {}
@@ -197,6 +198,10 @@ class DynamicEventTree(Grid):
     # Collect the branch info in a multi-level dictionary
     endInfo = {'endTime':self.actualEndTime,'endTimeStep':self.actualEndTs,'branchDist':list(self.actualBranchInfo.keys())[0]}
     endInfo['branchChangedParams'] = self.actualBranchInfo[endInfo['branchDist']]
+    # check if RELAP7 mode is activated, in case prepend the "<distribution>" string
+    if any("<distribution>" in s for s in self.branchProbabilities.keys()):
+      endInfo['branchDist'] = self.toBeSampled.keys()[self.toBeSampled.values().index(endInfo['branchDist'])]
+      #endInfo['branchDist'] = "<distribution>"+endInfo['branchDist']
     parentNode.add('actualEndTimeStep',self.actualEndTs)
     # # Get the parent element tree (xml object) to retrieve the information needed to create the new inputs
     # if(jobObject.identifier == self.TreeInfo[self.rootToJob[jobObject.identifier]].getrootnode().name): endInfo['parentNode'] = self.TreeInfo[self.rootToJob[jobObject.identifier]].getrootnode()
@@ -204,7 +209,10 @@ class DynamicEventTree(Grid):
     endInfo['parentNode'] = parentNode
     # get the branchedLevel dictionary
     branchedLevel = {}
-    for distk, distpb in zip(endInfo['parentNode'].get('initiatorDistribution'),endInfo['parentNode'].get('PbThreshold')): branchedLevel[distk] = utils.index(self.branchProbabilities[distk],distpb)
+    for distk, distpb in zip(endInfo['parentNode'].get('SampledVarsPb').keys(),endInfo['parentNode'].get('SampledVarsPb').values()):
+    #for distk, distpb in zip(endInfo['parentNode'].get('initiatorDistribution'),endInfo['parentNode'].get('PbThreshold')):
+      if distk not in self.epistemicVariables.keys(): branchedLevel[distk] = utils.index(self.branchProbabilities[distk],distpb)
+
     if not branchedLevel: self.raiseAnError(RuntimeError,'branchedLevel of node '+jobObject.identifier+'not found!')
     # Loop of the parameters that have been changed after a trigger gets activated
     for key in endInfo['branchChangedParams']:
@@ -297,7 +305,7 @@ class DynamicEventTree(Grid):
     # Check if endTime and endTimeStep (time step)  are present... In case store them in the relative working vars
     #try: #Branch info written out by program, so should always exist.
     self.actualEndTime = float(root.attrib['end_time'])
-    self.actualEndTs   = int(root.attrib['end_ts'])
+    self.actualEndTs   = int(root.attrib['end_ts']) if  'end_ts' in root.attrib.keys() else -1
     #except? pass
     # Store the information in a dictionary that has as keywords the distributions that triggered
     for node in root:
@@ -314,7 +322,7 @@ class DynamicEventTree(Grid):
       # we exit the loop here, because only one trigger at the time can be handled  right now
       break
     # remove the file
-    os.remove(filename)
+    if self.removeXmlBranchInfo: os.remove(filename)
     branchPresent = True
     return branchPresent
 
@@ -329,6 +337,9 @@ class DynamicEventTree(Grid):
       @ In, myInput, list, list of inputs for the Models object (passed through the Steps XML block)
       @ Out, None
     """
+    # add additional edits if needed
+    model.getAdditionalInputEdits(self.inputInfo)
+
     precSampled = rootTree.getrootnode().get('hybridsamplerCoordinate')
     rootnode    =  rootTree.getrootnode()
     rname       = rootnode.name
@@ -346,12 +357,17 @@ class DynamicEventTree(Grid):
     self.inputInfo['parentID'                  ] = 'root'
     self.inputInfo['conditionalPb'             ] = [1.0]
     self.inputInfo['conditionalPbr'            ] = 1.0
-    for key in self.branchProbabilities.keys():self.inputInfo['initiatorDistribution'].append(key.encode())
+    for key in self.branchProbabilities.keys():self.inputInfo['initiatorDistribution'].append(self.toBeSampled[key])
+    #for key in self.branchProbabilities.keys():self.inputInfo['initiatorDistribution'].append(key.encode())
     for key in self.branchProbabilities.keys():self.inputInfo['PbThreshold'].append(self.branchProbabilities[key][branchedLevel[key]])
+    #for key in self.branchProbabilities.keys():self.inputInfo['PbThreshold'].append(self.branchProbabilities[key][branchedLevel[key]])
     for key in self.branchProbabilities.keys():self.inputInfo['ValueThreshold'].append(self.branchValues[key][branchedLevel[key]])
+    #for key in self.branchProbabilities.keys():self.inputInfo['ValueThreshold'].append(self.branchValues[key][branchedLevel[key]])
     for varname in self.standardDETvariables:
-      self.inputInfo['SampledVars'  ][varname] = self.branchValues[self.toBeSampled[varname]][branchedLevel[self.toBeSampled[varname]]]
-      self.inputInfo['SampledVarsPb'][varname] = self.branchProbabilities[self.toBeSampled[varname]][branchedLevel[self.toBeSampled[varname]]]
+      self.inputInfo['SampledVars'  ][varname] = self.branchValues[varname][branchedLevel[varname]]
+      #self.inputInfo['SampledVars'  ][varname] = self.branchValues[self.toBeSampled[varname]][branchedLevel[self.toBeSampled[varname]]]
+      self.inputInfo['SampledVarsPb'][varname] = self.branchProbabilities[varname][branchedLevel[varname] ]
+      #self.inputInfo['SampledVarsPb'][varname] = self.branchProbabilities[self.toBeSampled[varname]][branchedLevel[self.toBeSampled[varname]]]
     if precSampled:
       for precSample in precSampled:
         self.inputInfo['SampledVars'  ].update(precSample['SampledVars'])
@@ -452,7 +468,8 @@ class DynamicEventTree(Grid):
       if self.branchCountOnLevel != 1: subGroup.add('conditionalPbr',condPbC)
       else: subGroup.add('conditionalPbr',condPbUn)
       # add initiator distribution info, start time, etc.
-      subGroup.add('initiatorDistribution',endInfo['branchDist'])
+
+      subGroup.add('initiatorDistribution',self.toBeSampled[endInfo['branchDist']])
       subGroup.add('startTime', endInfo['parentNode'].get('endTime'))
       # initialize the endTime to be equal to the start one... It will modified at the end of this branch
       subGroup.add('endTime', endInfo['parentNode'].get('endTime'))
@@ -473,6 +490,8 @@ class DynamicEventTree(Grid):
                 'conditionalPb':[subGroup.get('conditionalPbr')],
                 'startTime':endInfo['parentNode'].get('endTime'),
                 'parentID':subGroup.get('parent')}
+      # add additional edits if needed
+      model.getAdditionalInputEdits(self.inputInfo)
       # add the newer branch name to the map
       self.rootToJob[rname] = self.rootToJob[subGroup.get('parent')]
       # check if it is a preconditioned DET sampling, if so add the relative information
@@ -484,7 +503,7 @@ class DynamicEventTree(Grid):
       #  In this case there is not a probability threshold that needs to be added in the input
       #  for this particular distribution
       if not (branchedLevel[endInfo['branchDist']] >= len(self.branchProbabilities[endInfo['branchDist']])):
-        self.inputInfo['initiatorDistribution'] = [endInfo['branchDist']]
+        self.inputInfo['initiatorDistribution'] = [self.toBeSampled[endInfo['branchDist']]]
         self.inputInfo['PbThreshold'           ] = [self.branchProbabilities[endInfo['branchDist']][branchedLevel[endInfo['branchDist']]]]
         self.inputInfo['ValueThreshold'        ] = [self.branchValues[endInfo['branchDist']][branchedLevel[endInfo['branchDist']]]]
       #  For the other distributions, we put the unbranched thresholds
@@ -497,16 +516,18 @@ class DynamicEventTree(Grid):
         self.inputInfo['ValueThreshold'        ] = []
       # Add the unbranched thresholds
       for key in self.branchProbabilities.keys():
-        if not (key in endInfo['branchDist']) and (branchedLevel[key] < len(self.branchProbabilities[key])): self.inputInfo['initiatorDistribution'].append(key.encode())
+        if not (key in self.toBeSampled[endInfo['branchDist']]) and (branchedLevel[key] < len(self.branchProbabilities[key])): self.inputInfo['initiatorDistribution'].append(self.toBeSampled[key.encode()])
       for key in self.branchProbabilities.keys():
-        if not (key in endInfo['branchDist']) and (branchedLevel[key] < len(self.branchProbabilities[key])):
+        if not (key in self.toBeSampled[endInfo['branchDist']]) and (branchedLevel[key] < len(self.branchProbabilities[key])):
           self.inputInfo['PbThreshold'   ].append(self.branchProbabilities[key][branchedLevel[key]])
           self.inputInfo['ValueThreshold'].append(self.branchValues[key][branchedLevel[key]])
       self.inputInfo['SampledVars']   = {}
       self.inputInfo['SampledVarsPb'] = {}
       for varname in self.standardDETvariables:
-        self.inputInfo['SampledVars'][varname]   = self.branchValues[self.toBeSampled[varname]][branchedLevel[self.toBeSampled[varname]]]
-        self.inputInfo['SampledVarsPb'][varname] = self.branchProbabilities[self.toBeSampled[varname]][branchedLevel[self.toBeSampled[varname]]]
+        self.inputInfo['SampledVars'][varname]   = self.branchValues[varname][branchedLevel[varname]]
+        self.inputInfo['SampledVarsPb'][varname] = self.branchProbabilities[varname][branchedLevel[varname]]
+        #self.inputInfo['SampledVars'][varname]   = self.branchValues[self.toBeSampled[varname]][branchedLevel[self.toBeSampled[varname]]]
+        #self.inputInfo['SampledVarsPb'][varname] = self.branchProbabilities[self.toBeSampled[varname]][branchedLevel[self.toBeSampled[varname]]]
       if precSampled:
         for precSample in precSampled:
           self.inputInfo['SampledVars'  ].update(precSample['SampledVars'])
@@ -576,6 +597,7 @@ class DynamicEventTree(Grid):
     """
     #NB: if someday the DET handles restarts as other samplers do in generateInput, the return code 1 indicates the result
     #  is stored in a restart data object, while 0 indicates a new run has been found.
+    #model.getAdditionalInputEdits(self.inputInfo)
     return 0, self.localGenerateInput(model, oldInput)
 
   def localGenerateInput(self,model,myInput):
@@ -627,36 +649,64 @@ class DynamicEventTree(Grid):
       @ Out, None
     """
     Grid.localInputAndChecks(self,xmlNode)
-    if 'printEndXmlSummary' in xmlNode.attrib.keys():
-      if xmlNode.attrib['printEndXmlSummary'].lower() in utils.stringsThatMeanTrue(): self.printEndXmlSummary = True
-      else: self.printEndXmlSummary = False
-    if 'maxSimulationTime' in xmlNode.attrib.keys():
-      try:    self.maxSimulTime = float(xmlNode.attrib['maxSimulationTime'])
+    if 'printEndXmlSummary'  in xmlNode.attrib.keys(): self.printEndXmlSummary  = xmlNode.attrib['printEndXmlSummary'].lower()  in utils.stringsThatMeanTrue()
+    if 'removeXmlBranchInfo' in xmlNode.attrib.keys(): self.removeXmlBranchInfo = xmlNode.attrib['removeXmlBranchInfo'].lower() in utils.stringsThatMeanTrue()
+    if 'maxSimulationTime'   in xmlNode.attrib.keys():
+      try                        : self.maxSimulTime = float(xmlNode.attrib['maxSimulationTime'])
       except (KeyError,NameError): self.raiseAnError(IOError,'Can not convert maxSimulationTime in float number!!!')
     branchedLevel, error_found = {}, False
-    gridInfo = self.gridEntity.returnParameter("gridInfo")
+    gridInfo   = self.gridEntity.returnParameter("gridInfo")
     errorFound = False
+    errorMsgs  = ''
+
     for keyk in self.axisName:
-      branchedLevel[self.toBeSampled[keyk]] = 0
+      branchedLevel[keyk] = 0
+      #branchedLevel[self.toBeSampled[keyk]] = 0
       self.standardDETvariables.append(keyk)
       if self.gridInfo[keyk] == 'CDF':
-        self.branchProbabilities[self.toBeSampled[keyk]] = gridInfo[keyk][2]
-        self.branchProbabilities[self.toBeSampled[keyk]].sort(key=float)
-        if max(self.branchProbabilities[self.toBeSampled[keyk]]) > 1:
-          self.raiseAWarning("One of the Thresholds for distribution " + str(gridInfo[keyk][2]) + " is > 1")
+        self.branchProbabilities[keyk] = gridInfo[keyk][2]
+        self.branchProbabilities[keyk].sort(key=float)
+        if max(self.branchProbabilities[keyk]) > 1:
+          errorMsgs += "One of the Thresholds for distribution " + str(gridInfo[keyk][2]) + " is > 1 \n"
           errorFound = True
-          for index in range(len(sorted(self.branchProbabilities[self.toBeSampled[keyk]], key=float))):
-            if sorted(self.branchProbabilities[self.toBeSampled[keyk]], key=float).count(sorted(self.branchProbabilities[self.toBeSampled[keyk]], key=float)[index]) > 1:
-              self.raiseAWarning("In distribution " + str(self.toBeSampled[keyk]) + " the Threshold " + str(sorted(self.branchProbabilities[self.toBeSampled[keyk]], key=float)[index])+" appears multiple times!!")
-              errorFound = True
+        probMultiplicities = Counter(self.branchProbabilities[keyk])
+        multiples = [prob for prob,mult in probMultiplicities.items() if mult > 1]
+        ## Only the multiple variables remain
+        for prob in multiples:
+          errorMsgs += "In variable " + str(keyk) + " the Threshold " + str(prob)+" appears multiple times!!\n"
+          errorFound = True
+#         self.branchProbabilities[self.toBeSampled[keyk]] = gridInfo[keyk][2]
+#         self.branchProbabilities[self.toBeSampled[keyk]].sort(key=float)
+#         if max(self.branchProbabilities[self.toBeSampled[keyk]]) > 1:
+#           self.raiseAWarning("One of the Thresholds for distribution " + str(gridInfo[keyk][2]) + " is > 1")
+#           errorFound = True
+#           for index in range(len(sorted(self.branchProbabilities[self.toBeSampled[keyk]], key=float))):
+#             if sorted(self.branchProbabilities[self.toBeSampled[keyk]], key=float).count(sorted(self.branchProbabilities[self.toBeSampled[keyk]], key=float)[index]) > 1:
+#               self.raiseAWarning("In distribution " + str(self.toBeSampled[keyk]) + " the Threshold " + str(sorted(self.branchProbabilities[self.toBeSampled[keyk]], key=float)[index])+" appears multiple times!!")
+#               errorFound = True
       else:
-        self.branchValues[self.toBeSampled[keyk]] = gridInfo[keyk][2]
-        self.branchValues[self.toBeSampled[keyk]].sort(key=float)
-        for index in range(len(sorted(self.branchValues[self.toBeSampled[keyk]], key=float))):
-          if sorted(self.branchValues[self.toBeSampled[keyk]], key=float).count(sorted(self.branchValues[self.toBeSampled[keyk]], key=float)[index]) > 1:
-            self.raiseAWarning("In distribution " + str(self.toBeSampled[keyk]) + " the Threshold " + str(sorted(self.branchValues[self.toBeSampled[keyk]], key=float)[index])+" appears multiple times!!")
-            errorFound = True
-    if errorFound: self.raiseAnError(IOError,"In sampler named " + self.name+' Errors have been found!' )
+        self.branchValues[keyk] = gridInfo[keyk][2]
+        self.branchValues[keyk].sort(key=float)
+        valueMultiplicities = Counter(self.branchValues[keyk])
+        multiples = [value for value,mult in valueMultiplicities.items() if mult > 1]
+        ## Only the multiple variables remain
+        for value in multiples:
+          errorMsgs += "In variable " + str(keyk) + " the Threshold " + str(value)+" appears multiple times!!\n"
+          errorFound = True
+#         self.branchValues[self.toBeSampled[keyk]] = gridInfo[keyk][2]
+#         self.branchValues[self.toBeSampled[keyk]].sort(key=float)
+#         for index in range(len(sorted(self.branchValues[self.toBeSampled[keyk]], key=float))):
+#           if sorted(self.branchValues[self.toBeSampled[keyk]], key=float).count(sorted(self.branchValues[self.toBeSampled[keyk]], key=float)[index]) > 1:
+#             self.raiseAWarning("In distribution " + str(self.toBeSampled[keyk]) + " the Threshold " + str(sorted(self.branchValues[self.toBeSampled[keyk]], key=float)[index])+" appears multiple times!!")
+#             errorFound = True
+
+    # check if RELAP7 mode is activated, in case check that a <distribution> variable is unique in the input
+    if any("<distribution>" in s for s in self.branchProbabilities.keys()):
+      associatedDists = self.toBeSampled.values()
+      if len(list(set(associatedDists))) != len(associatedDists):
+        errorMsgs += "Distribution-mode sampling activated in " + self.name+". In this case every <distribution> needs to be assocaited with one single <Distribution> block!\n"
+        errorFound = True
+    if errorFound: self.raiseAnError(IOError,"In sampler named " + self.name+' the following errors have been found: \n'+errorMsgs )
     # Append the branchedLevel dictionary in the proper list
     self.branchedLevel.append(branchedLevel)
 
@@ -692,9 +742,9 @@ class DynamicEventTree(Grid):
     """
     paramDict = {}
     for key in self.branchProbabilities.keys():
-      paramDict['Probability Thresholds for dist ' + str(key) + ' are: '] = [str(x) for x in self.branchProbabilities[key]]
+      paramDict['Probability Thresholds for var ' + str(key) + ' are: '] = [str(x) for x in self.branchProbabilities[key]]
     for key in self.branchValues.keys()       :
-      paramDict['Values Thresholds for dist ' + str(key) + ' are: '] = [str(x) for x in self.branchValues[key]]
+      paramDict['Values Thresholds for var ' + str(key) + ' are: '] = [str(x) for x in self.branchValues[key]]
     return paramDict
 
   def localGetCurrentSetting(self):
@@ -759,8 +809,10 @@ class DynamicEventTree(Grid):
 
     for key in self.branchProbabilities.keys():
       #kk = self.toBeSampled.values().index(key)
-      self.branchValues[key] = [self.distDict[self.toBeSampled.keys()[self.toBeSampled.values().index(key)]].ppf(float(self.branchProbabilities[key][index])) for index in range(len(self.branchProbabilities[key]))]
+      #self.branchValues[key] = [self.distDict[self.toBeSampled.keys()[self.toBeSampled.values().index(key)]].ppf(float(self.branchProbabilities[key][index])) for index in range(len(self.branchProbabilities[key]))]
+      self.branchValues[key] = [self.distDict[key].ppf(float(self.branchProbabilities[key][index])) for index in range(len(self.branchProbabilities[key]))]
     for key in self.branchValues.keys():
       #kk = self.toBeSampled.values().index(key)
-      self.branchProbabilities[key] = [self.distDict[self.toBeSampled.keys()[self.toBeSampled.values().index(key)]].cdf(float(self.branchValues[key][index])) for index in range(len(self.branchValues[key]))]
+      #self.branchProbabilities[key] = [self.distDict[self.toBeSampled.keys()[self.toBeSampled.values().index(key)]].cdf(float(self.branchValues[key][index])) for index in range(len(self.branchValues[key]))]
+      self.branchProbabilities[key] = [self.distDict[key].cdf(float(self.branchValues[key][index])) for index in range(len(self.branchValues[key]))]
     self.limit = sys.maxsize
